rpc channel - 2

+++ rocket/common/msg_id_util.cc
@@ -30,7 +30,7 @@ std::string MsgIDUtil::GenMsgID() {
     }
     t_msg_id_no = res;
   } else {
-    int i = t_msg_id_no.length() - 1;
+    size_t i = t_msg_id_no.length() - 1;
     while (t_msg_id_no[i] == '9' && i >= 0) {
       i--;
     }
